+++ lib/deployex/release/gcp_storage.ex
@@ -5,8 +5,8 @@ defmodule Deployex.Release.GcpStorage do
 
   @behaviour Deployex.Release.Adapter
 
+  alias Deployex.Catalog
   alias Deployex.Status
-  alias Deployex.Storage
   alias Deployex.Upgrade
 
   require Logger
@@ -21,7 +21,7 @@ defmodule Deployex.Release.GcpStorage do
   @impl true
   def get_current_version_map do
     path =
-      "https://storage.googleapis.com/#{bucket()}/versions/#{Storage.monitored_app_name()}/#{env()}/current.json"
+      "https://storage.googleapis.com/#{bucket()}/versions/#{Catalog.monitored_app_name()}/#{env()}/current.json"
 
     :get
     |> Finch.build(path, headers(), [])
@@ -39,8 +39,8 @@ defmodule Deployex.Release.GcpStorage do
   def download_and_unpack(instance, version) do
     {:ok, download_path} = Briefly.create()
 
-    app_name = Storage.monitored_app_name()
-    app_lang = Storage.monitored_app_lang()
+    app_name = Catalog.monitored_app_name()
+    app_lang = Catalog.monitored_app_lang()
 
     gcp_path =
       "https://storage.googleapis.com/#{bucket()}/dist/#{app_name}/#{app_name}-#{version}.tar.gz"
@@ -57,7 +57,7 @@ defmodule Deployex.Release.GcpStorage do
     end
 
     Status.clear_new(instance)
-    new_path = Storage.new_path(instance)
+    new_path = Catalog.new_path(instance)
 
     {"", 0} = System.cmd("tar", ["-x", "-f", download_path, "-C", new_path])