+++ lib/deployex/release/s3.ex
@@ -5,8 +5,8 @@ defmodule Deployex.Release.S3 do
 
   @behaviour Deployex.Release.Adapter
 
+  alias Deployex.Catalog
   alias Deployex.Status
-  alias Deployex.Storage
   alias Deployex.Upgrade
 
   require Logger
@@ -20,7 +20,7 @@ defmodule Deployex.Release.S3 do
   """
   @impl true
   def get_current_version_map do
-    path = "versions/#{Storage.monitored_app_name()}/#{env()}/current.json"
+    path = "versions/#{Catalog.monitored_app_name()}/#{env()}/current.json"
 
     bucket()
     |> ExAws.S3.get_object(path)
@@ -38,8 +38,8 @@ defmodule Deployex.Release.S3 do
   def download_and_unpack(instance, version) do
     {:ok, download_path} = Briefly.create()
 
-    app_name = Storage.monitored_app_name()
-    app_lang = Storage.monitored_app_lang()
+    app_name = Catalog.monitored_app_name()
+    app_lang = Catalog.monitored_app_lang()
 
     s3_path = "dist/#{app_name}/#{app_name}-#{version}.tar.gz"
 
@@ -49,7 +49,7 @@ defmodule Deployex.Release.S3 do
       |> ExAws.request()
 
     Status.clear_new(instance)
-    new_path = Storage.new_path(instance)
+    new_path = Catalog.new_path(instance)
 
     {"", 0} = System.cmd("tar", ["-x", "-f", download_path, "-C", new_path])