+++ lib/deployex_web/live/applications/logs.ex
@@ -96,13 +96,13 @@ defmodule DeployexWeb.ApplicationsLive.Logs do
   defp log_path(instance, :logs_stdout) do
     instance
     |> String.to_integer()
-    |> Deployex.Storage.stdout_path()
+    |> Deployex.Catalog.stdout_path()
   end
 
   defp log_path(instance, :logs_stderr) do
     instance
     |> String.to_integer()
-    |> Deployex.Storage.stderr_path()
+    |> Deployex.Catalog.stderr_path()
   end
 
   defp tail_if_exists(%{assigns: %{id: id, action: action}} = socket, path) do