+++ test/deployex/catalog_test.exs
@@ -1,0 +1,5 @@
+defmodule Deployex.CatalogTest do
+  use ExUnit.Case, async: true
+
+  doctest Deployex.Catalog
+end