+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Enhancements
  * [[`PR-100`](https://github.com/thiagoesteves/deployex/pull/100)] Adding new log server to persist logs within ets tables
+ * [[`PR-101`](https://github.com/thiagoesteves/deployex/pull/101)] Rename Deployex.Storage to Deployex.Catalog
 
 ## 0.3.2 🚀 (2025-03-21)